Skip "Email Change" email for Shibboleth users
WordPress, by default, notifies users when their email has changed because the user, by default, uses that email address for authentication purposes. When the email field is managed by the Shibboleth plugin, it will update the user account to match the fields coming from Shibboleth attributes. Given that the purpose of the notification is not relevant to Shibboleth authenticated users, it makes sense to not send the notification.